fastapi-service/pyproject.toml
[project]
name = "fastapi-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.21.0",
    "aiosqlite>=0.21.0",
    "fastapi[all]>=0.116.1",
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
    "rq>=2.5.0",
    "sqlalchemy>=2.0.43",
]
